Prevent caching issues
There's some odd behaviour with webpack-offline-plugin in how it handles
/ and /index.html routes. For some reason it ignores /index.html and
always fetches the actual file from the server. On / route the behaviour
is correct; the HTML is cached by the service worker and properly
updated after a deployment.

sw-precache didn't have this issue so it appears to be a bug in the
webpack-offline-plugin.
For now instead of working around it we simply redirect /index.html to
/.